==> scripts/01_import and describe.ipynb <==
# Contents List

01. Import libraries
02. Import and explore data
    a. Orders data
    b. Products data
03. Answer questions with data


# 01. Import libraries

# import libraries
import pandas as pd
import numpy as np
import os

# 02. Import and explore data

# Orders data

# create shortcut for data imports
path = r'C:\Users\jacym\Desktop\Career Foundry projects\04-2023 Instacart basket analysis'

# create vars_list shortcut
vars_list = ['order_id', 'user_id', 'order_number', 'order_dow', 'order_hour_of_day', 'days_since_prior_order']

# import data with vars_list shortcut
df = pd.read_csv(os.path.join(path, '02 Data', 'Original data', 'orders.csv'), usecols = vars_list)

# print first 5 rows of df
df.head()

# Products data

# import products.csv
df_prods = pd.read_csv(os.path.join(path, '02 Data', 'Original data', 'products.csv'), index_col = False)

#print first 20 rows
df_prods.head(20)

# print last 35 rows
df_prods.tail(35)

# print column names
df_prods.columns

# print number of rows, columns
df_prods.shape

# 03. Answering questions with data

# display descriptive statistics
df_prods.describe()

Max value for aisle_id: 134

# display data types
df_prods.dtypes

department_id data type: integer

df.shape
