==> scripts/02_ wrangle and subset.ipynb <==
# Contents List

01. Import libraries
02. Import data
03. Data wrangling exercise
04. Data wrangling task

# 01. Import libraries

# import libraries
import pandas as pd
import numpy as np
import os

# 02. Import data

# create shortcut for data imports
path = r'C:\Users\jacym\Desktop\Career Foundry projects\04-2023 Instacart basket analysis'

# import orders.csv
df_ords = pd.read_csv(os.path.join(path, '02 Data', 'Original data', 'orders.csv'), index_col = False)

# import products.csv
df_prods = pd.read_csv(os.path.join(path, '02 Data', 'Original data', 'products.csv'), index_col = False)

# import departments.csv
df_dep = pd.read_csv(os.path.join(path, '02 Data', 'Original data', 'departments.csv'), index_col = False)

# 03. Data wrangling exercise

# drop eval_set column
df_ords = df_ords.drop(columns = 'eval_set')

# count blanks in days_since_prior_order columns
df_ords['days_since_prior_order'].value_counts(dropna = False)

# rename column
df_ords.rename(columns = {'order_dow' : 'order_day_of_week'}, inplace = True)

df_ords.head()

# change data type for order_id to string so describe() disregards it
df_ords['order_id'] = df_ords['order_id'].astype('str')

# check that it worked
df_ords['order_id'].dtype

df_dep.head()

#transpose df (switch rows/columns)
df_dep.T

# overwrite df with transposed df
df_dep_t = df_dep.T

# add index
df_dep_t.reset_index()

# create new header
new_header = df_dep_t.iloc[0]

# create new dataset without first row
df_dep_t_new = df_dep_t[1:]

# set column names with new_header variable established earlier
df_dep_t_new.columns = new_header

df_dep_t_new.head()

# turn df_dep into data dictionary
data_dict = df_dep_t_new.to_dict('index')

data_dict

# view first 5 rows
df_prods.head()

# retrieve value from data dictionary
print(data_dict.get('19'))

# produce df that lists snacks - items with dept id 19
df_snacks = df_prods[df_prods['department_id'] == 19]

# view first 10 rows of snacks df
df_snacks.head(10)

# 04. Data wrangling task

df_ords.head()

# Step 2 - change data type

# change user_id to string because the numbers can't be usefully analyzed
df_ords['user_id'] = df_ords['user_id'].astype('str')

#check that it worked. O = success!
df_ords['user_id'].dtype

# Step 3 - rename column w/out overwriting

# attempt to rename user_id column as customer_id without overwriting data frame
df_ords_renamed = df_ords.rename(columns = {'user_id' : 'customer_id'}, inplace = False)

# check alternate df
df_ords_renamed.head()

# Step 4 - busiest hour for orders

# display frequency counts of hour_of_day column
df_ords['order_hour_of_day'].value_counts(dropna = False)

Answer- Hour 10

# Step 5- use dictionary to find dept name

print(data_dict.get('4'))

# Step 6- pull up breakfast sales

# view data dictionary
data_dict

# create subset of breakfast items
df_breakfast = df_prods[df_prods['department_id'] == 14]

# Step 7- create multi-dept subset

df_dinner_parties = df_prods.loc[df_prods['department_id'].isin([5,20,7,12])]

# Step 8- count rows in dataframe

df_dinner_parties

Answer: 7650 rows

# Step 9- extract info about user 1

# create subset for user 1
df_user_1 = df_ords.loc[df_ords['user_id'] == 1]

#view subset
df_user_1

# Step 10- View stats about subset

# view descriptive stats of subset
df_user_1.describe()

# Step 12- Export orders data

df_ords.to_csv(r'C:\Users\jacym\Desktop\Career Foundry projects\04-2023 Instacart basket analysis\02 Data\Prepared data\orders_wrangled.csv', index=False)

# Step 13- Export departments data

df_dep_t_new.to_csv(os.path.join(path, '02 Data','Prepared Data','departments_wrangled.csv'))
