==> scripts/03_consistency checks.ipynb <==
# Contents List

01. Import libraries
02. Import data
03. Exercise- Addressing mixed-type data
04. Exercise- Finding missing data
05. Exercise- Addressing missing data
06. Exercise- Finding duplicates
07. Excercise- Export data

Task
08. Describe data
09. Check for mixed data types
10. Check for missing values
11. Address missing values
12. Find duplicates
13. Address duplicates
14. Investigate price anomalies
15. Export checked data

# 01. Import libraries

# import libraries
import pandas as pd
import numpy as np
import os

# 02. Import data

# create shortcut for data imports
path = r'C:\Users\jacym\Desktop\Career Foundry projects\04-2023 Instacart basket analysis'

# import products.csv
df_prods = pd.read_csv(os.path.join(path, '02 Data', 'Original data', 'products.csv'), index_col = False)

# import orders_wrangled.csv
df_ords = pd.read_csv(os.path.join(path, '02 Data', 'Prepared data', 'orders_wrangled.csv'), index_col = False)

df_ords.describe()

# 03. Exercise- Addressing mixed-type data

# Create sample dataframe
df_test = pd.DataFrame()

# create a new column and fill it with varying data types
df_test['mix'] = 'a', 'b', 1, True

df_test.head()

# check df for mixed types- will print any mixed columns
for col in df_test.columns.tolist():
  weird = (df_test[[col]].applymap(type) != df_test[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_test[weird]) > 0:
    print (col)

# identify the appropriate data type for the column, then run this to change the type of any non-conforming values
df_test['mix'] = df_test['mix'].astype('str')

# 04. Exercise- Finding missing data

# to find missing values
df_prods.isnull().sum()

# create df subset where product_name is blank
df_nan = df_prods[df_prods['product_name'].isnull() == True]

df_nan

# 05. Exercise- Addressing missing data

# taking delete approach, but you can also impute mean/median or create flag variable if appropriate
# count rows in original df
df_prods.shape

# create new df without blanks
df_prods_clean = df_prods[df_prods['product_name'].isnull() == False]

# verify that rows have decreased by number of missing values
df_prods_clean.shape

# 06. Exercise- Finding duplicates

# create subset to view duplicates
df_dups = df_prods_clean[df_prods_clean.duplicated()]

df_dups

df_prods_clean.shape

# create new df with duplicates removed
df_prods_clean_no_dups = df_prods_clean.drop_duplicates()

df_prods_clean_no_dups.shape

# 07. Exercise- Export data

df_prods_clean_no_dups.to_csv(r'C:\Users\jacym\Desktop\Career Foundry projects\04-2023 Instacart basket analysis\02 Data\Prepared data\products_checked.csv', index=False)

# Task

# 08. Describe data

df_ords.describe()

The count is lower for the days_since_prior_order column, which I'm assuming is because of blanks. The max seems a bit high for order_number. Otherwise everything looks OK so far.

# view rows with order number over 99
df_ords[df_ords['order_number']> 99.0]

On review, it looks like order_number maxes out at 100 and there are 1,374 customers who've reached that threshold.

# 09. Check for mixed data types

# check df for mixed types- will print any mixed columns
for col in df_ords.columns.tolist():
  weird = (df_ords[[col]].applymap(type) != df_ords[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_ords[weird]) > 0:
    print (col)

# 10. Check for missing values

# find missing values
df_ords.isnull().sum()

# create df subset where days_since_prior_order is blank
df_ords_nan = df_ords[df_ords['days_since_prior_order'].isnull() == True]

df_ords_nan.describe()

There are 206,209 rows where the days_since_prior_order column is blank. For all of those entries, the order was the customer's first order.

# 11. Address missing values

The rows with missing days_since_prior_order values represent useful data about first-time customers, so I don't want to remove them or compromise the data by imputing. I think the best approach would be to add a binary column such as first_time_customer that acts as a flag. I don't know how to do that yet, so for now I'm creating two new, separate dataframes with and without the blanks which I'll call df_ords_nan and df_ords_returning (for returning customers).

df_ords.shape

df_ords_returning = df_ords[df_ords['days_since_prior_order'].isnull() == False]

df_ords_returning.shape

# 12. Find duplicates

# create subset to view duplicates
df_ords_dups = df_ords[df_ords.duplicated()]

df_ords_dups

I didn't find any duplicates, so it appears each row is unique.

# 13. Address duplicates

I didn't find any duplicates, so not taking any corrective measures here.

# 14. Investigate price anomalies

df_prods_clean_no_dups.describe()

Max price is too high

#check for other high prices
df_prods_clean_no_dups[df_prods_clean_no_dups['prices']> 30.0]

#correct inflated prices
df_prods_clean_no_dups = df_prods_clean_no_dups.replace ({'prices':{99999.0: .99, 14900.0:1.49 }})

df_prods_clean_no_dups.describe()

# 15. Export checked data

df_ords.to_csv(r'C:\Users\jacym\Desktop\Career Foundry projects\04-2023 Instacart basket analysis\02 Data\Prepared data\all_orders_checked.csv', index=False)

df_ords_returning.to_csv(r'C:\Users\jacym\Desktop\Career Foundry projects\04-2023 Instacart basket analysis\02 Data\Prepared data\returning_orders_checked.csv', index=False)

df_ords_nan.to_csv(r'C:\Users\jacym\Desktop\Career Foundry projects\04-2023 Instacart basket analysis\02 Data\Prepared data\first_time_orders_checked.csv', index=False)
