==> scripts/04_combine data step 1.ipynb <==
# Contents List

01. Import libraries
02. Import and inspect data
03. Merge dataframes
04. Export merge as pkl

# 01. Import Libraries

# import libraries
import pandas as pd
import numpy as np
import os

# 02. Import and inspect data

# create shortcut for data imports
path = r'C:\Users\jacym\Desktop\Career Foundry projects\04-2023 Instacart basket analysis'

# import checked orders data
df_ords = pd.read_csv(os.path.join(path, '02 Data', 'Prepared data', 'all_orders_checked.csv'), index_col = False)

# import new data - order_products_prior
df_ords_prior = pd.read_csv(os.path.join(path, '02 Data', 'Original data', 'order_products_prior.csv'), index_col = False)

# consistency checks- blanks
df_ords_prior.isnull().sum()

# check for duplicates
df_ords_prior_dups = df_ords_prior[df_ords_prior.duplicated()]

# check for duplicates
df_ords_prior_dups

# check summary statistics
df_ords_prior.describe()

# check out orders with high add_to_cart_order values
df_ords_prior[df_ords_prior['add_to_cart_order']> 50.0]

# Note: It looks like this column represents the order in which the item was added to the virtual cart, so the high values are only present for very large orders. Check out one large order to confirm...

df_ords_prior[df_ords_prior['order_id'] == 1171]

# Confirmed- column value increases as customer adds items.

# check output
df_ords_prior.head()

df_ords_prior.shape

df_ords.shape

df_ords.head()

# 03. Merge dataframes

df_merged_large = df_ords.merge(df_ords_prior, on = 'order_id', indicator = True)

df_merged_large.head()

df_merged_large['_merge'].value_counts()

df_merged_large.shape

# 04. Export data as pkl

# export data to pkl

df_merged_large.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'orders_products_combined.pkl'))
