==> scripts/05_combine data step 2.ipynb <==
# Contents list

01. Import libraries
02. Import data
03. Merge data
04. Export merge as pkl

# 01. Import libraries

# import libraries
import pandas as pd
import numpy as np
import os

# 02. Import merged data and product data

# create shortcut for data imports
path = r'C:\Users\jacym\Desktop\Career Foundry projects\04-2023 Instacart basket analysis'

# import merged data
df_orders_products_merged = pd.read_pickle(os.path.join(path, '02 Data','Prepared data', 'orders_products_combined.pkl'))

# import products data
df_products_list = pd.read_csv(os.path.join(path, '02 Data', 'Prepared data', 'products_checked.csv'), index_col = False)

# check shape- should be same as exported data set
df_orders_products_merged.shape

# check output
df_orders_products_merged.head()

# check shape of products data
df_products_list.shape

df_products_list.head()

# 03. Merge data

# drop indicator column from merged data - you can't have two merge flag columns
df_orders_products_merged = df_orders_products_merged.drop(columns = ['_merge'])

# use inner join to merge dataframes on product_id column (be wary of memory errors, was able to do this after closing all running scripts and cleaning up code)
df_all_merged_large = df_orders_products_merged.merge(df_products_list, on = 'product_id', indicator = True)

df_all_merged_large['_merge'].value_counts()

# take a peek at your merged data
df_all_merged_large.head()

# 04. Export merged data as pkl

# export data as pkl bc it's a really big file
df_all_merged_large.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'orders_products_merged.pkl'))
