==> scripts/06_ derive variables.ipynb <==
# Contents List

01. Import libraries
02. Import and narrow data
03. Define and execute user-derived function
04. If-statements with loc()
05. Execute loc() on whole df
06. If-statements with for-loops
07. Task

# 01. Import libraries

# import libraries
import pandas as pd
import numpy as np
import os

# 02. Import and narrow data

# create shortcut for data imports
path = r'C:\Users\jacym\Desktop\Career Foundry projects\04-2023 Instacart basket analysis'

# import merged data
df_orders_products_merged = pd.read_pickle(os.path.join(path, '02 Data','Prepared data', 'orders_products_merged.pkl'))

# narrow dataframe to first million rows to prevent memory errors while working with user-derived functions
df = df_orders_products_merged[:1000000]

# check work
df.shape

# 03. Define and execute user-derived function

# define new price_label function

def price_label(row):

  if row['prices'] <= 5:
    return 'Low-range product'
  elif (row['prices'] > 5) and (row['prices'] <= 15):
    return 'Mid-range product'
  elif row['prices'] > 15:
    return 'High range'
  else: return 'Not enough data'

# execute function -- in a new column called price_range, execute the price_label function on each row (each column would be axis=0)
df['price_range'] = df.apply(price_label, axis=1)
# red warning message is not an error

# run value_counts function to see how labels are distributed. Note there are no high-range items in this subset of the data
df['price_range'].value_counts()

# confirm the value count by looking for the max price. If it's over 15, there's an error in your code
df['prices'].max()

# 04. If-statements with loc()
As warning message above suggests, the loc() function (a pre-defined function rather than user-derived) is a better way to handle these types of operations. Here, if-then logic is implied.

# first set conditions, each in its own cell
# if price is more than 15, then print high-range product in new column called price_range_loc
df.loc[df['prices'] > 15, 'price_range_loc'] = 'High-range product'

# if price is between 5 and 15, then print mid-range product (separate arguments -- <= 15 and > 5 -- with parentheses)
df.loc[(df['prices'] <= 15) & (df['prices'] > 5), 'price_range_loc'] = 'Mid-range product'

# if price is less than 5, then print low-range product
df.loc[df['prices'] <= 5, 'price_range_loc'] = 'Low-range product'

# check value_counts -- same as above, but this way should be faster and yield no warning message
df['price_range_loc'].value_counts()

# 05. Execute loc() on whole df
Since loc() is faster, we can use it on the complete df

# if price is more than 15, then print high-range product in new column called price_range_loc
df_orders_products_merged.loc[df_orders_products_merged['prices'] > 15, 'price_range_loc'] = 'High-range product'

# if price is between 5 and 15, then print mid-range product (separate arguments -- <= 15 and > 5 -- with parentheses)
df_orders_products_merged.loc[(df_orders_products_merged['prices'] <=15) & (df_orders_products_merged['prices'] > 5), 'price_range_loc'] = 'Mid-range product'

# if price is less than 5, then print low-range product
df_orders_products_merged.loc[df_orders_products_merged['prices'] <= 5, 'price_range_loc'] = 'Low-range product'

# check value counts on full data frame. Now we have some high-range entries.
df_orders_products_merged['price_range_loc'].value_counts()

df_orders_products_merged.head()

# 06. If-statements with for-loops
This is faster than a user-derived function bc you're only looping through one column

Goal: Create new column displaying business of order day

# run frequency check
df_orders_products_merged['order_day_of_week'].value_counts()

# create empty list, then run through day of week column and print one of the following if each condition is met

result = []

for value in df_orders_products_merged['order_day_of_week']:
  if value == 0:
    result.append('Busiest day')
  elif value == 4:
    result.append('Least busy')
  else:
    result.append('Regularly busy')

# view list
result

# insert list in df by creating new column and filling it with your result list
df_orders_products_merged['busiest_day'] = result

# run frequency check of newly created column to see orders by business category (you can cross-check this with frequency check from 19)
df_orders_products_merged['busiest_day'].value_counts(dropna = False)

# 07. Task

Step 2: Create busiest_days column

# create new list 'result2' for new classification of day business
# must use 'or' as seen below to append multiple values to a condition. 
result2 = []

for value in df_orders_products_merged['order_day_of_week']:
  if value == 0 or value == 1:
    result2.append('Busy day')
  elif value == 4 or value == 3:
    result2.append('Slow day')
  else:
    result2.append('Regular day')

# create new column and fill with result2
df_orders_products_merged['busiest_days'] = result2

Step 3: Check frequency of new column

# run frequency check of new column to see how many orders were logged in each category
df_orders_products_merged['busiest_days'].value_counts()

Observations: The highest number of orders are logged during the three regular days of the week (Monday, Thursday and Friday), followed closely by the busiest days (Saturday and Sunday). The slow days of Tuesday and Wednesday account for considerably fewer orders.

Step 4: Create busiest_period_of_day column

# check frequency of hour of day column to see the slowest and busiest periods
df_orders_products_merged['order_hour_of_day'].value_counts()

# Create for-loop designating hours 9-16 as most orders, 0-6 as fewest orders, and everything else as average
hours = []

for value in df_orders_products_merged['order_hour_of_day']:
  if value >= 9 and value <= 16:
    hours.append('Most orders')
  elif value >= 0 and value <= 6:
    hours.append('Fewest orders')
  else:
    hours.append('Average orders')

# create new column with hours results
df_orders_products_merged['busiest_period_of_day'] = hours

Step 5: Check frequency of new column

df_orders_products_merged['busiest_period_of_day'].value_counts()

Step 7: Export data with derivations as pkl

df_orders_products_merged.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'orders_products_merged_expanded.pkl'))
