==> scripts/08_ prepare data for merge.ipynb <==
# Contents List

01. Import and check data
02. Rename columns
03. Drop columns
04. Data consistency checks
05. Export wrangled data

# 01. Import and check data

# import libraries
import pandas as pd
import numpy as np
import os

# create shortcut for data imports
path = r'C:\Users\jacym\Desktop\Career Foundry projects\04-2023 Instacart basket analysis'

# import customer data
df_customers = pd.read_csv(os.path.join(path, '02 Data','Original data', 'customers.csv'))

df_customers.head()

df_customers.shape

# 02. Rename columns for clarity and consistency

# rename columns - all lower case, use underscores
df_customers.rename(columns = {'First Name' : 'first_name' , 'Surnam' : 'last_name' , 'Gender' : 'gender' , 'STATE' : 'state' , 'Age' : 'age'}, inplace = True)

df_customers.head()

# view descriptive statistics
df_customers.describe()

# check tail 
df_customers.tail(20)

df_customers['date_joined'].value_counts(dropna = False)

# 03. Drop unneeded columns

Observations: Descriptive stats look normal so far. In order to protect user privacy, I'll create a new df without customer names. User IDs suffice as unique identifiers.

# drop name columns from df
df_customers_anon = df_customers.drop(columns = ['first_name', 'last_name'])

# check work
df_customers_anon.head()

# 04. Data consistency checks

# view data types
df_customers_anon.dtypes

# convert user id to object so describe disregards it
df_customers_anon['user_id'] = df_customers_anon['user_id'].astype('str')

# check work
df_customers_anon.describe()

# check df for mixed types- will print any mixed columns
for col in df_customers_anon.columns.tolist():
  weird = (df_customers_anon[[col]].applymap(type) != df_customers_anon[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_customers_anon[weird]) > 0:
    print (col)

Observation: No columns printed = no mixed data types

# check frequency for fam status (just curious)
df_customers_anon['fam_status'].value_counts(dropna = False)

# check frequency for state (these results would be a red flag if the data were real)
df_customers_anon['state'].value_counts(dropna = False)

df_customers_anon.isnull().sum()

Observation: No blanks

# create subset to check for duplicates
df_dups = df_customers_anon[df_customers_anon.duplicated()]

# view duplicates
df_dups

Observation: No duplicates

# checking to make sure very high incomes are rare in the data -- 119 of 206,209 rows, checks out
df_customers_anon[df_customers_anon['income']> 450000]

Conclusions: The data seems to be in order. I only needed to eliminate name columns for privacy and change the data type of user_id.

# 05. Export wrangled data

df_customers_anon.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'customers_wrangled.pkl'))
