==> scripts/09_merge data.ipynb <==
# Contents List

01. Import libraries and data
02. Check data types
03. Merge data
04. Export merged data

# 01. Import libraries and data

# import libraries
import pandas as pd
import numpy as np
import os

# create shortcut for data imports
path = r'C:\Users\jacym\Desktop\Career Foundry projects\04-2023 Instacart basket analysis'

# import customer data
df_customers = pd.read_pickle(os.path.join(path, '02 Data','Prepared data', 'customers_wrangled.pkl'))

# import orders data
df_orders = pd.read_pickle(os.path.join(path, '02 Data','Prepared data', 'orders_products_merged_grouped.pkl'))

# 02. Check data types

df_orders.dtypes

# drop extra merge column
df_orders = df_orders.drop(columns = '_merge')

# 03. Merge data

orders_customers_merged = df_orders.merge(df_customers, on = 'user_id', indicator = True)

orders_customers_merged['_merge'].value_counts()

orders_customers_merged.head()

# drop merge column
orders_customers_merged = orders_customers_merged.drop(columns = '_merge')

# 04. Export merged data

# export merged data
orders_customers_merged.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'orders_customers_merged.pkl'))
