==> scripts/11_crosstab consistency check.ipynb <==
# Contents List

01. Import libraries and data
02. Create crosstab 
03. Copy crosstab to Excel

# 01. Import libraries and data

# import libraries
import pandas as pd
import numpy as np
import os

# create shortcut for data imports
path = r'C:\Users\jacym\Desktop\Career Foundry projects\04-2023 Instacart basket analysis'

# import merged data
ords_prods_merge = pd.read_pickle(os.path.join(path, '02 Data','Prepared data', 'orders_products_merged_expanded.pkl'))

# 02. Create crosstab

# create crosstab for days since prior order and order number to see where missing data is concentrated
crosstab = pd.crosstab(ords_prods_merge['days_since_prior_order'], ords_prods_merge['order_number'], dropna = False)

# 03. Copy crosstab to Excel

# copy crosstab to view in excel -- 0s runing down the order 1 column indicates missing values are for first orders
crosstab.to_clipboard()
