==> scripts/12_create new category.ipynb <==
# Contents List

01. Import libraries and data
02. Address security implications
03a. Create regional segmentation of data

# 01. Import libraries and data

# import libraries
import pandas as pd
import numpy as np
import os

# create shortcut for data imports
path = r'C:\Users\jacym\Desktop\Career Foundry projects\04-2023 Instacart basket analysis'

# import data
df = pd.read_pickle(r'C:\Users\jacym\Desktop\Career Foundry projects\04-2023 Instacart basket analysis\02 Data\Prepared data\orders_customers_merged.pkl')

# check data
df.shape

# 02. Address security implications
I addressed privacy issues in task 4.9a by dropping the first and last names columns from the customer data set and exporting the anonymized version. The customer names are not included in the merged data.

# 03a. Create a regional segmentation of the data

# create a list called regions containing the region for each state
regions = {
    
    'Massachusetts': 'Northeast',
    'Maine': 'Northeast',
    'New York': 'Northeast',
    'Rhode Island': 'Northeast',
    'Connecticut': 'Northeast',
    'New Hampshire': 'Northeast',    
    'New Jersey': 'Northeast',   
    'Vermont' : 'Northeast',
    'Pennsylvania' : 'Northeast',     
    'Wisconsin': 'Midwest',
    'Michigan': 'Midwest',
    'Ohio': 'Midwest',
    'North Dakota': 'Midwest',
    'South Dakota': 'Midwest',
    'Nebraska': 'Midwest',
    'Kansas': 'Midwest',
    'Iowa': 'Midwest',
    'Missouri': 'Midwest',
    'Illinois': 'Midwest',
    'Indiana': 'Midwest',
    'Minnesota': 'Midwest',
    'Washington': 'West',
    'Oregon': 'West',
    'California': 'West', 
    'Idaho': 'West',
    'Montana': 'West',
    'Wyoming': 'West',
    'Nevada': 'West',
    'Utah': 'West',
    'Colorado': 'West',
    'Arizona': 'West',
    'New Mexico': 'West',
    'Alaska': 'West',
    'Hawaii': 'West',
    'Delaware': 'South',
    'Maryland': 'South',
    'District of Columbia': 'South',
    'Virginia': 'South',
    'West Virginia': 'South',
    'North Carolina': 'South',
    'South Carolina': 'South',
    'Georgia': 'South',
    'Florida': 'South',
    'Kentucky': 'South',
    'Tennessee': 'South',
    'Mississippi': 'South',
    'Alabama': 'South',
    'Oklahoma': 'South',
    'Texas': 'South',
    'Arkansas': 'South',
    'Louisiana': 'South'
}

# define new function get_region- return region from list associated with value in state column
def get_region(state):
    return regions[state]

# create a new column called region and fill it by applying result of my get_region function
df['region'] = df['state'].apply(get_region)

# check work
df.head()

df.shape

# view user ID totals grouped by region
df.groupby(['region']).agg({'user_id': ['nunique']})

# view spending_flag grouped by region
df.groupby(['region', 'spending_flag']).agg({'spending_flag': ['count']})

Observations: About 99.62% of customers classify as low spenders in the Midwest, South and West regions. 
The percentage is slightly higher in the Northeast, where 99.67% of customers classify as low spenders.

# export df with regions
df.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'orders_customers_regions.pkl'))
