==> scripts/13_extract subset.ipynb <==
# Contents List

(continued from 4.10 part 1a Regions)
04. Create exclusion flag for low-activity customers

# import libraries
import pandas as pd
import numpy as np
import os

# create shortcut for data imports
path = r'C:\Users\jacym\Desktop\Career Foundry projects\04-2023 Instacart basket analysis'

# import data
df = pd.read_pickle(r'C:\Users\jacym\Desktop\Career Foundry projects\04-2023 Instacart basket analysis\02 Data\Prepared data\orders_customers_regions.pkl')

# 04. Create exclusion flag for low-activity customers

# loc: create exclusion flag with two labels, each in separate cell
# label 1: active
df.loc[df['max_order'] >= 5, 'low_activity_flag'] = 'Active customer'

# label 2: inactive
df.loc[df['max_order'] < 5, 'low_activity_flag'] = 'Inactive customer'

# check
df.head()

# check frequency on new exclusion flag column
df['low_activity_flag'].value_counts()

# create new version of df dropping all rows of inactive customers
df_active = df[df.low_activity_flag != 'Inactive customer']

# rows in active df match active customer count, so the drop was successful
df_active.shape

# export active df
df_active.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'orders_customers_merged_active.pkl'))
