==> scripts/16_depts by day analysis.ipynb <==
# Contents List

01. Import libraries and data
02. View products ordered per department by day

# 01. Import libraries and data

# import libraries
import pandas as pd
import numpy as np
import os
import matplotlib.pyplot as plt
import seaborn as sns
import scipy

# create shortcut for data imports
path = r'C:\Users\jacym\Desktop\Career Foundry projects\04-2023 Instacart basket analysis'

# import merged data
df = pd.read_pickle(os.path.join(path, '02 Data','Prepared data', 'orders_products_merged_grouped.pkl'))

# 02. View products ordered per department by day

# groupby day and dept id, agg dept_id count
df.groupby(['order_day_of_week', 'department_id']).agg({'department_id': ['count']})

pd.options.display.max_rows = 150

df.groupby(['department_id']).agg({'department_id': ['count']})
