==> scripts/17_downscaling.ipynb <==
# Contents List

01. Import libraries and data
02. View dataframe types and memory usage
03. Drop unneeded columns
04. View memory usage again
05. Export downscaled data

# import libraries
import pandas as pd
import numpy as np
import os

# create shortcut for data imports
path = r'C:\Users\jacym\Desktop\Career Foundry projects\04-2023 Instacart basket analysis'

# import data
df = pd.read_pickle(r'C:\Users\jacym\Desktop\Career Foundry projects\04-2023 Instacart basket analysis\02 Data\Prepared data\orders_customers_regions.pkl')

# view memory usage
df.info(memory_usage = 'deep')

# drop unneeded columns
df = df.drop(columns = ['reordered', 'aisle_id', 'price_range_loc', 'busiest_days'])

# view memory usage
df.info(memory_usage = 'deep')

# export downscaled data
df.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'orders_customers_regions.pkl'))
