==> scripts/18_additional downscaling.ipynb <==
# Contents List

01. Import libraries and data
02. View dataframe types and memory usage
03. Run downcasting loop
04. View memory usage again
05. Drop unneeded columns

# import libraries
import pandas as pd
import numpy as np
import os

# create shortcut for data imports
path = r'C:\Users\jacym\Desktop\Career Foundry projects\04-2023 Instacart basket analysis'

# import merged data
df = pd.read_pickle(os.path.join(path, '02 Data','Prepared data', 'orders_products_merged_grouped.pkl'))

# view memory usage
df.info(memory_usage = 'deep')

# downcasting loop
for column in df:
 if df[column].dtype == 'float64':
     df[column]=pd.to_numeric(df[column], downcast='float')
 if df[column].dtype == 'int64':
     df[column]=pd.to_numeric(df[column], downcast='integer')


# view memory usage again
df.info(memory_usage = 'deep')

# drop unneeded column
df = df.drop(columns = 'busiest_day')

# change data type
df['user_id'] = df['user_id'].astype('str')

# change data type
df['order_id'] = df['order_id'].astype('str')

# export
df.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'orders_products_merged_grouped.pkl'))
